# docs/swagger.yaml
definitions:
  MenuPayloads.ArticleBodyDetail:
    properties:
      article_body_paragraph:
        type: string
      article_image_content_path:
        type: string
    type: object
  MenuPayloads.ArticleInsertPayloads:
    properties:
      article_body_paragraph:
        items:
          $ref: '#/definitions/MenuPayloads.ArticleBodyDetail'
        type: array
      article_header:
        description: ArticleId                int                     `json:"article_id"`
        type: string
      article_header_path_content:
        type: string
    type: object
...
  MenuPayloads.ArticleUpdatePayloads:
    properties:
      article_body_content:
        items:
          $ref: '#/definitions/MenuPayloads.ArticleBodyDetail'
        type: array
      article_id:
        type: integer
    type: object
...
  MenuPayloads.ProfilePayloadRequest:
    properties:
      user_email:
        type: string
      user_gender:
        type: string
      user_height:
        type: number
      user_name:
        type: string
      user_phone_number:
        type: string
      user_profile_description:
        type: string
      user_profile_image:
        type: string
      user_weight:
        description: |-
          UserDetailId           int     `gorm:"column:user_detail_id;primaryKey;not null" json:"user_detail_id"`
          UserId                 int     `gorm:"column:user_id" json:"user_id"`
        type: number
    type: object
...
  entities.ArticleBodyEntities:
    properties:
      article_body_id:
        type: integer
      article_body_paragraph:
        type: string
      article_id:
        type: integer
      article_image_content_path:
        type: string
    type: object
  entities.ArticleEntities:
    properties:
      article_body:
        items:
          $ref: '#/definitions/entities.ArticleBodyEntities'
        type: array
      article_created_by_user_id:
        type: integer
      article_date_created:
        type: string
      article_header:
        type: string
      article_header_path_content:
        type: string
      article_id:
        type: integer
    type: object
...
  responses.ErrorResponses:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
  responses.StandarAPIResponses:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: devin@gmail.com
    name: IoT Blynk Smart Agriculture Backend Service
    url: IoT
  description: IoT Blynk Smart Agriculture Backend Service
  license:
    name: IoT Blynk Smart Agriculture Backend Service
  termsOfService: http://swagger.io/terms/
  title: IoT Blynk Smart Agriculture Backend Service
  version: "1.0"
paths:
  /api/article:
    get:
      consumes:
      - application/json
      description: Get All Article By Pagination
      parameters:
      - description: sort_by
        in: query
        name: sort_by
        type: string
      - description: sort_of
        in: query
        name: sort_of
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.ArticleEntities'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Get All Article By Pagination
      tags:
      - Article
    patch:
      consumes:
      - application/json
      description: Update Article
      parameters:
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MenuPayloads.ArticleUpdatePayloads'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Update Article
      tags:
      - Article
    post:
      consumes:
      - application/json
      description: Create New Article
      parameters:
      - description: Insert Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MenuPayloads.ArticleInsertPayloads'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.StandarAPIResponses'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Create New Article
      tags:
      - Article
  /api/article/by-id/{article_id}:
    get:
      consumes:
      - application/json
      description: Get Article By Article
      parameters:
      - description: article_id
        in: path
        name: article_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ArticleEntities'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Get Article By Article
      tags:
      - Article
  /api/article/delete/{article_id}:
    delete:
      consumes:
      - application/json
      description: Delete Article
      parameters:
      - description: article_id
        in: path
        name: article_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.StandarAPIResponses'
      security:
      - BearerAuth: []
      summary: Delete Article
      tags:
      - Article
  /api/article/search:
    get:
      consumes:
      - application/json
      description: Get All Article By Pagination With Filter
      parameters:
      - description: key_filter
        in: query
        name: key_filter
        type: string
      - description: sort_by
        in: query
        name: sort_by
        type: string
      - description: sort_of
        in: query
        name: sort_of
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.ArticleEntities'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Get All Article By Pagination With Filter
      tags:
      - Article
...
  /api/profile:
    patch:
      consumes:
      - application/json
      description: Get Profile Detail
      parameters:
      - description: user detail Headers
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MenuPayloads.ProfilePayloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Get Profile Detail
      tags:
      - Profile
    post:
      consumes:
      - application/json
      description: Get Profile Detail
      parameters:
      - description: user detail Headers
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MenuPayloads.ProfilePayloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Get Profile Detail
      tags:
      - Profile
  /api/profile/:
    get:
      consumes:
      - application/json
      description: Get Profile Detail
      parameters:
      - description: user_id
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Get Profile Detail
      tags:
      - Profile
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"